+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - Bugfix: MUC commands were being ignored
 - UI: Always show the OMEMO lock icon (grayed out if not available).
+- #1353 Message Delivery Receipts not working because of the message "type" attribute
 - #1374 Can't load embedded chat when changing `view_mode` between page reloads
 - #1376 Fixed some alignment issues in the sidebar
 - #1378 Message Delivery Receipts were being sent for carbons and own messages